Moved client unit tests back to the root upon request to avoid merge conflicts.
