+++ services/editor/spec/unit/components/JPadFullEditor/rules-utils-spec.js
@@ -1,8 +1,8 @@
 /* global jest, before, beforeEach, describe, it, expect */
 import { assert, expect } from 'chai';
-import * as RulesUtils from '../../../../../src/components/JPadFullEditor/rules-utils';
+import * as RulesUtils from '../../../../src/components/JPadFullEditor/rules-utils';
 
-jest.unmock('../../../../../src/components/JPadFullEditor/rules-utils');
+jest.unmock('../../../../src/components/JPadFullEditor/rules-utils');
 
 const toExplicitRule = (Value, ...matchers) => ({
   Matcher: matchers.reduce((result, matcher) => ({ ...result, [matcher]: 'matcherValue' }), {}),
