+++ services/editor/spec/unit/utils/key-name-validations-spec.js
@@ -1,10 +1,10 @@
 /* global jest, beforeEach, describe, it, expect */
-jest.unmock('../../../../src/pages/keys/components/KeyPage/KeyAddPage/key-name-validations');
+jest.unmock('../../../src/pages/keys/components/KeyPage/KeyAddPage/key-name-validations');
 jest.unmock('chance');
 
 import { assert, expect } from 'chai';
-import keyNameValidations from '../../../../src/pages/keys/components/KeyPage/KeyAddPage/key-name-validations';
-import { BLANK_KEY_NAME } from '../../../../src/store/ducks/ducks-utils/blankKeyDefinition';
+import keyNameValidations from '../../../src/pages/keys/components/KeyPage/KeyAddPage/key-name-validations';
+import { BLANK_KEY_NAME } from '../../../src/store/ducks/ducks-utils/blankKeyDefinition';
 
 describe('key-name-validations', () => {
   const categoryName1 = 'someCategoryName1';
